+++ CHANGELOG.md
@@ -5,11 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+Nothing yet
+
+## [0.2.11] - 2025-04-05
+### Fixed
+- Fix Rails 8 integration (#189)
+
+## [0.2.10] - 2025-03-17 (yanked)
 ### Added
 - Support for Ruby 3.4 (#186)
 ### Removed
 - Drop Ruby 3.0 support (#186)
 - Drop Rails 6.1 support (#186)
+### Fixed
+- Add `text_area` alias to `textarea` helper to follow Rails 8.0 change (#181)
 
 ## [0.2.9] - 2024-11-09
 ### Added
@@ -128,7 +137,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add CHANGELOG (#50)
 - Add CI (#2)
 
-[Unreleased]: https://github.com/pantographe/view_component-form/compare/v0.2.9...HEAD
+[Unreleased]: https://github.com/pantographe/view_component-form/compare/v0.2.11...HEAD
+[0.2.11]: https://github.com/pantographe/view_component-form/compare/v0.2.10...v0.2.11
+[0.2.10]: https://github.com/pantographe/view_component-form/compare/v0.2.9...v0.2.10
 [0.2.9]: https://github.com/pantographe/view_component-form/compare/v0.2.8...v0.2.9
 [0.2.8]: https://github.com/pantographe/view_component-form/compare/v0.2.7...v0.2.8
 [0.2.7]: https://github.com/pantographe/view_component-form/compare/v0.2.6...v0.2.7